// front/src/app/enums/interests.ts
export enum Interest {
    NAT = "Nature et paysages (montagnes, plages, parcs naturels, lacs, etc.)",
    CUL = "Culture (musées, monuments, histoire, art, etc.)",
    GAS = "Gastronomie (spécialités locales, vins, restaurants, etc.)",
    VNO = "Vie nocturne (bars, clubs, concerts, festivals, etc.)",
    SPT = "Sports et activités de plein air (randonnée, vélo, surf, escalade, ski, etc.)",
    DET = "Bien-être et détente (spa, thermes, yoga, massages, etc.)",
    SHO = "Shopping (marchés, boutiques, centres commerciaux, etc.)",
    EVS = "Événements spéciaux (carnavals, foires, festivals, etc.)",
    FAM = "Famille et enfants (parcs d'attractions, zoos, aquariums, etc.)",
    ARC ="Architecture et design (bâtiments historiques, modernes, art urbain, etc.)",
}